Fixed typecasting errors (typecasting 64-bit pointer to 32-bit int)

+++ src/gravity/gravity_sensor.cpp
@@ -127,7 +127,7 @@ bool gravity_sensor::on_start(void)
 	AUTOLOCK(m_mutex);
 
 	m_orientation_sensor->add_client(ORIENTATION_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_orientation_sensor->add_interval((int)this, (m_interval/MS_TO_US), true);
+	m_orientation_sensor->add_interval((long int)this, (m_interval/MS_TO_US), true);
 	m_orientation_sensor->start();
 
 	activate();
@@ -139,14 +139,14 @@ bool gravity_sensor::on_stop(void)
 	AUTOLOCK(m_mutex);
 
 	m_orientation_sensor->delete_client(ORIENTATION_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_orientation_sensor->delete_interval((int)this, true);
+	m_orientation_sensor->delete_interval((long int)this, true);
 	m_orientation_sensor->stop();
 
 	deactivate();
 	return true;
 }
 
-bool gravity_sensor::add_interval(int client_id, unsigned int interval)
+bool gravity_sensor::add_interval(long int client_id, unsigned int interval)
 {
 	AUTOLOCK(m_mutex);
 	m_orientation_sensor->add_interval(client_id , interval, true);
@@ -154,7 +154,7 @@ bool gravity_sensor::add_interval(int client_id, unsigned int interval)
 	return sensor_base::add_interval(client_id, interval, true);
 }
 
-bool gravity_sensor::delete_interval(int client_id)
+bool gravity_sensor::delete_interval(long int client_id)
 {
 	AUTOLOCK(m_mutex);
 	m_orientation_sensor->delete_interval(client_id , true);
