+++ src/orientation/orientation_sensor.cpp
@@ -235,13 +235,13 @@ bool orientation_sensor::on_start(void)
 	AUTOLOCK(m_mutex);
 
 	m_accel_sensor->add_client(ACCELEROMETER_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_accel_sensor->add_interval((int)this, (m_interval/MS_TO_US), true);
+	m_accel_sensor->add_interval((long int)this, (m_interval/MS_TO_US), true);
 	m_accel_sensor->start();
 	m_gyro_sensor->add_client(GYROSCOPE_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_gyro_sensor->add_interval((int)this, (m_interval/MS_TO_US), true);
+	m_gyro_sensor->add_interval((long int)this, (m_interval/MS_TO_US), true);
 	m_gyro_sensor->start();
 	m_magnetic_sensor->add_client(GEOMAGNETIC_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_magnetic_sensor->add_interval((int)this, (m_interval/MS_TO_US), true);
+	m_magnetic_sensor->add_interval((long int)this, (m_interval/MS_TO_US), true);
 	m_magnetic_sensor->start();
 
 	activate();
@@ -253,20 +253,20 @@ bool orientation_sensor::on_stop(void)
 	AUTOLOCK(m_mutex);
 
 	m_accel_sensor->delete_client(ACCELEROMETER_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_accel_sensor->delete_interval((int)this, true);
+	m_accel_sensor->delete_interval((long int)this, true);
 	m_accel_sensor->stop();
 	m_gyro_sensor->delete_client(GYROSCOPE_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_gyro_sensor->delete_interval((int)this, true);
+	m_gyro_sensor->delete_interval((long int)this, true);
 	m_gyro_sensor->stop();
 	m_magnetic_sensor->delete_client(GEOMAGNETIC_EVENT_RAW_DATA_REPORT_ON_TIME);
-	m_magnetic_sensor->delete_interval((int)this, true);
+	m_magnetic_sensor->delete_interval((long int)this, true);
 	m_magnetic_sensor->stop();
 
 	deactivate();
 	return true;
 }
 
-bool orientation_sensor::add_interval(int client_id, unsigned int interval)
+bool orientation_sensor::add_interval(long int client_id, unsigned int interval)
 {
 	AUTOLOCK(m_mutex);
 
@@ -277,7 +277,7 @@ bool orientation_sensor::add_interval(int client_id, unsigned int interval)
 	return sensor_base::add_interval(client_id, interval, true);
 }
 
-bool orientation_sensor::delete_interval(int client_id)
+bool orientation_sensor::delete_interval(long int client_id)
 {
 	AUTOLOCK(m_mutex);
